Change auth_miss_mongodb from virtualization layer to app layer
